refactor: Buyer를 Prizes로 바꾸고 메서드 리팩토링

+++ src/main/java/lotto/domain/LottoMachine.java
@@ -10,7 +10,7 @@ import java.util.stream.Stream;
 
 public class LottoMachine {
 
-    public static final int LOTTO_PRICE = 1000;
+    public static final long LOTTO_PRICE = 1000;
 
     public List<Lotto> issue(int money) {
         validateMoney(money);
